% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adjusted_cost_base.R
\name{calculate_adjusted_cost_base}
\alias{calculate_adjusted_cost_base}
\title{calculate adjusted cost base}
\usage{
calculate_adjusted_cost_base(df)
}
\arguments{
\item{df}{}
}
\value{
a table containing the original df augmented with the share_balance, adjusted_cost_base and capital_gain columns
}
\description{
calculate adjusted cost base
}
\examples{
library(acb)
calculate_adjusted_cost_base(df_example)
}
