% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_questrade_csv.R
\name{import_questrade_csv}
\alias{import_questrade_csv}
\title{import_questrade_csv
Import a questrade trade confirmation format and manipulate the columns
for compatibility with the calculate_adjusted_cost_base() function
an example csv is included in the package for comparison.}
\usage{
import_questrade_csv(path)
}
\arguments{
\item{path}{path to the questrade trade confirmations csv}
}
\value{
a tibble containing an imported csv that was
}
\description{
import_questrade_csv
Import a questrade trade confirmation format and manipulate the columns
for compatibility with the calculate_adjusted_cost_base() function
an example csv is included in the package for comparison.
}
\examples{
library(acb)
df_questrade <- acb::import_questrade_csv(system.file("extdata", "questrade_trade_confirmations.csv", package="acb"))
calculate_adjusted_cost_base(df_questrade)
}
